backend/requirements.txt
fastapi
numpy
opencv-python
pillow
python-multipart
uvicorn
onnxruntime
facenet_pytorch 
torch
torchvision
mmcv